Shake off more unused types (and meta-types)
"Meta" types bring confusion to the table, nothing else. Name vars properly. You won't needa alias a string.
