+++ src/shared/components/Balance/index.tsx
@@ -1,12 +1,11 @@
 import { FunctionComponent, ReactNode } from 'react'
 import React, { Fragment } from 'react'
 import cx from 'classnames'
-import { NumberString } from '~/shared/types/common-types'
 import styles from './balance.pcss'
 
 type AccountProps = {
     name: string
-    value: NumberString
+    value: string
 }
 
 export const Account: FunctionComponent<AccountProps> = ({
