+++ src/shared/components/Ui/Text/FlushHistoryDecorator.tsx
@@ -1,6 +1,6 @@
 import { ComponentType } from 'react'
 import React, { useCallback, useState, forwardRef } from 'react'
-import '~/shared/types/common-types'
+
 export type Props = {
     flushHistoryOnBlur?: boolean
     onBlur?: ((arg0: React.FocusEvent<EventTarget>) => void) | null | undefined
