+++ src/shared/components/Ui/Text/RevertOnEscapeDecorator.tsx
@@ -1,6 +1,5 @@
 import { ChangeEvent, ComponentType } from 'react'
 import React, { useState, forwardRef, useEffect, useCallback } from 'react'
-import '~/shared/types/common-types'
 export type Props = {
     onChange?: ((arg0: React.SyntheticEvent<EventTarget>) => void) | null | undefined
     onKeyDown?: ((arg0: React.KeyboardEvent<EventTarget>) => void) | null | undefined
